+++ test/units/map-response.test.ts
@@ -300,7 +300,10 @@ describe('Map Response', () => {
 		expect(response).toBeInstanceOf(Response)
 		expect(await response.text()).toEqual('Hina')
 		expect(response.headers.get('name')).toEqual('Sorasaki Hina')
-		expect(response.headers.getAll('set-cookie')).toEqual(['name=hina', 'affiliation=gehenna'])
+		expect(response.headers.getAll('set-cookie')).toEqual([
+			'name=hina',
+			'affiliation=gehenna'
+		])
 	})
 
 	it('map Passthrough', async () => {
@@ -308,6 +311,19 @@ describe('Map Response', () => {
 
 		expect(response).toBeInstanceOf(Response)
 		expect(await response.text()).toEqual('hi')
+		expect(response.status).toBe(200)
+	})
+
+	it('map video content-range', async () => {
+		const kyuukararin = Bun.file('test/kyuukurarin.mp4')
+
+		const response = mapResponse(kyuukararin, defaultContext)
+
+		expect(response).toBeInstanceOf(Response)
+		expect(response.headers.get('accept-ranges')).toEqual('bytes')
+		expect(response.headers.get('content-range')).toEqual(
+			`bytes 0-${kyuukararin.size - 1}/${kyuukararin.size}`
+		)
 		expect(response.status).toBe(200)
 	})
 })